depthai/requirements.txt
imutils==0.5.3
concurrent-log-handler==0.9.16
opencv-python==4.5.1.48
depthai==2.13.3.0
blobconverter==1.2.8
pyserial==3.4